Refactor audit command helpers into submodules
