+++ crates/copybook-cli/src/commands/audit.rs
@@ -625,486 +625,22 @@ pub async fn run(
     }
 }
 
-#[allow(
-    clippy::too_many_arguments,
-    clippy::fn_params_excessive_bools,
-    clippy::used_underscore_binding
-)]
-#[derive(Deserialize)]
-struct RawAccessEvent {
-    #[serde(default)]
-    user: Option<String>,
-    #[serde(default)]
-    user_id: Option<String>,
-    #[serde(default)]
-    action: Option<String>,
-    #[serde(default)]
-    access_type: Option<String>,
-    #[serde(default)]
-    resource_type: Option<String>,
-    #[serde(default)]
-    resource: Option<String>,
-    #[serde(default)]
-    resource_id: Option<String>,
-    #[serde(default)]
-    result: Option<String>,
-    #[serde(default)]
-    status: Option<String>,
-    #[serde(default)]
-    source_ip: Option<String>,
-    #[serde(default)]
-    user_agent: Option<String>,
-    #[serde(default)]
-    source_ip_address: Option<String>,
-}
+mod access_log;
+mod health;
+mod schema_analysis;
+mod support;
 
-#[derive(Deserialize)]
-struct RawHealthEvent {
-    #[serde(default)]
-    event_id: Option<String>,
-    #[serde(default)]
-    timestamp: Option<String>,
-    #[serde(default)]
-    source: Option<String>,
-    #[serde(default)]
-    event_type: Option<String>,
-    #[serde(default)]
-    integrity_hash: Option<String>,
-    #[serde(default)]
-    previous_hash: Option<String>,
-}
-
-#[derive(Serialize, Deserialize)]
-struct HealthEventRecord {
-    event_id: String,
-    timestamp: String,
-    source: String,
-    event_type: String,
-    integrity_hash: String,
-    previous_hash: Option<String>,
-}
-
-#[derive(Deserialize)]
-struct ReportSectionFile {
-    #[serde(default)]
-    compliance_validation: Option<Value>,
-    #[serde(default)]
-    lineage_analysis: Option<Value>,
-    #[serde(default)]
-    performance_audit: Option<Value>,
-    #[serde(default)]
-    security_audit: Option<Value>,
-}
-
-fn parse_health_events(path: &Path) -> AuditResult<(Vec<HealthEventRecord>, Vec<String>)> {
-    let content = fs::read_to_string(path)?;
-    let mut parse_issues = Vec::new();
-
-    if content.trim().is_empty() {
-        return Ok((Vec::new(), parse_issues));
-    }
-
-    if let Ok(mut events) = serde_json::from_str::<Vec<RawHealthEvent>>(&content) {
-        let mut parsed = Vec::with_capacity(events.len());
-        for (index, raw_event) in events.drain(..).enumerate() {
-            match normalize_health_event(raw_event) {
-                Ok(event) => parsed.push(event),
-                Err(err) => parse_issues.push(format!(
-                    "entry {}: failed to normalize audit health event: {err}",
-                    index + 1
-                )),
-            }
-        }
-        return Ok((parsed, parse_issues));
-    }
-
-    let mut parsed = Vec::new();
-    for (line_number, raw_line) in content.lines().enumerate() {
-        let line = raw_line.trim();
-        if line.is_empty() {
-            continue;
-        }
-
-        match serde_json::from_str::<RawHealthEvent>(line) {
-            Ok(raw_event) => match normalize_health_event(raw_event) {
-                Ok(event) => parsed.push(event),
-                Err(err) => parse_issues.push(format!(
-                    "line {}: failed to normalize audit health event: {err}",
-                    line_number + 1
-                )),
-            },
-            Err(err) => parse_issues.push(format!(
-                "line {}: invalid health event JSON: {err}",
-                line_number + 1
-            )),
-        }
-    }
-
-    Ok((parsed, parse_issues))
-}
-
-fn parse_audit_events_for_health(path: &Path) -> AuditResult<(Vec<AuditEvent>, Vec<String>)> {
-    let content = fs::read_to_string(path)?;
-    let mut parse_issues = Vec::new();
-
-    if content.trim().is_empty() {
-        return Ok((Vec::new(), parse_issues));
-    }
-
-    if let Ok(events) = serde_json::from_str::<Vec<AuditEvent>>(&content) {
-        return Ok((events, parse_issues));
-    }
-
-    let mut events = Vec::new();
-    for (line_number, raw_line) in content.lines().enumerate() {
-        let line = raw_line.trim();
-        if line.is_empty() {
-            continue;
-        }
-
-        match serde_json::from_str::<AuditEvent>(line) {
-            Ok(event) => events.push(event),
-            Err(err) => parse_issues.push(format!(
-                "line {}: not a parseable AuditEvent JSON object: {err}",
-                line_number + 1
-            )),
-        }
-    }
-
-    Ok((events, parse_issues))
-}
-
-fn normalize_health_event(raw: RawHealthEvent) -> Result<HealthEventRecord, String> {
-    let event_id = raw
-        .event_id
-        .unwrap_or_else(|| format!("health-event-{}", generate_random_suffix()));
-    let timestamp = raw
-        .timestamp
-        .unwrap_or_else(|| chrono::Utc::now().to_rfc3339());
-    let source = raw.source.unwrap_or_else(|| "copybook-core".to_string());
-    let event_type = raw.event_type.unwrap_or_else(|| "Unknown".to_string());
-    let integrity_hash = raw.integrity_hash.unwrap_or_else(|| "".to_string());
-
-    Ok(HealthEventRecord {
-        event_id,
-        timestamp,
-        source,
-        event_type,
-        integrity_hash,
-        previous_hash: raw.previous_hash,
-    })
-}
-
-fn generate_random_suffix() -> String {
-    chrono::Utc::now()
-        .timestamp_nanos_opt()
-        .map_or_else(|| "fallback".to_string(), |value| value.to_string())
-}
-
-#[inline]
-fn bytes_to_gbps(bytes_per_second: f64) -> f64 {
-    bytes_per_second / (1024.0_f64 * 1024.0_f64 * 1024.0_f64)
-}
-
-#[inline]
-fn bytes_to_mbps(bytes_per_second: f64) -> f64 {
-    bytes_per_second / (1024.0_f64 * 1024.0_f64)
-}
-
-fn combine_exit_code(current: ExitCode, candidate: ExitCode) -> ExitCode {
-    if candidate.precedence() >= current.precedence() {
-        candidate
-    } else {
-        current
-    }
-}
-
-fn sidecar_path(output: &Path, suffix: &str) -> PathBuf {
-    let stem = output
-        .file_stem()
-        .and_then(|s| s.to_str())
-        .unwrap_or("audit");
-    let ext = output
-        .extension()
-        .and_then(|s| s.to_str())
-        .unwrap_or("json");
-    let file_name = format!("{stem}.{suffix}.{ext}");
-    output
-        .parent()
-        .unwrap_or_else(|| Path::new("."))
-        .join(file_name)
-}
-
-fn parse_compliance_profile(raw: &str) -> Option<ComplianceProfile> {
-    match raw.to_ascii_lowercase().as_str() {
-        "sox" => Some(ComplianceProfile::SOX),
-        "hipaa" => Some(ComplianceProfile::HIPAA),
-        "gdpr" => Some(ComplianceProfile::GDPR),
-        "pci" | "pcidss" | "pci_dss" | "pci-dss" => Some(ComplianceProfile::PciDss),
-        _ => None,
-    }
-}
-
-fn parse_compliance_profiles(
-    profiles: &str,
-) -> Result<(Vec<ComplianceProfile>, Vec<String>), io::Error> {
-    let mut parsed = Vec::new();
-    let mut names = Vec::new();
-    let mut seen = HashSet::new();
-
-    for raw in profiles
-        .split(',')
-        .map(str::trim)
-        .filter(|item| !item.is_empty())
-    {
-        if let Some(profile) = parse_compliance_profile(raw) {
-            if seen.insert(profile) {
-                names.push(raw.to_string());
-                parsed.push(profile);
-            }
-        } else {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                format!("invalid compliance profile: '{raw}'"),
-            ));
-        }
-    }
-
-    Ok((parsed, names))
-}
-
-fn parse_copybook_schema(path: &Path) -> AuditResult<Schema> {
-    let copybook_text = fs::read_to_string(path)?;
-    Ok(copybook_core::parse_copybook(&copybook_text)?)
-}
-
-fn collect_leaf_fields<'a>(fields: &'a [Field], out: &mut Vec<&'a Field>) {
-    for field in fields {
-        if field.children.is_empty() {
-            out.push(field);
-        } else {
-            collect_leaf_fields(&field.children, out);
-        }
-    }
-}
-
-fn map_field_type_name(field: &Field) -> String {
-    match &field.kind {
-        copybook_core::FieldKind::Alphanum { len } => format!("alphanum[{len}]"),
-        copybook_core::FieldKind::ZonedDecimal { digits, .. } => format!("zoned-decimal[{digits}]"),
-        copybook_core::FieldKind::BinaryInt { bits, .. } => format!("binary-int[{bits}]"),
-        copybook_core::FieldKind::PackedDecimal { digits, .. } => {
-            format!("packed-decimal[{digits}]")
-        }
-        copybook_core::FieldKind::Group => "group".to_string(),
-        copybook_core::FieldKind::Condition { .. } => "condition".to_string(),
-        copybook_core::FieldKind::Renames { .. } => "renames".to_string(),
-        copybook_core::FieldKind::EditedNumeric { width, .. } => format!("edited-numeric[{width}]"),
-        copybook_core::FieldKind::FloatSingle => "float-single".to_string(),
-        copybook_core::FieldKind::FloatDouble => "float-double".to_string(),
-    }
-}
-
-fn estimate_schema_bytes(schema: &Schema) -> (u64, u64, u64) {
-    let mut display_bytes = 0u64;
-    let mut comp3_bytes = 0u64;
-    let mut total_bytes = 0u64;
-
-    let mut fields = Vec::new();
-    collect_leaf_fields(&schema.fields, &mut fields);
-    for field in fields {
-        let (display, comp3) = match &field.kind {
-            FieldKind::Alphanum { len } => (*len as u64, 0),
-            FieldKind::PackedDecimal { digits, .. } => (0, u64::from(*digits) / 2 + 1),
-            FieldKind::BinaryInt { bits, .. } => (0, u64::from(*bits) / 8),
-            FieldKind::ZonedDecimal { digits, .. } => (0, (u64::from(*digits) + 1) / 2),
-            FieldKind::EditedNumeric { width, .. } => (*width as u64, 0),
-            FieldKind::FloatSingle => (0, 4),
-            FieldKind::FloatDouble => (0, 8),
-            _ => (0, 0),
-        };
-
-        display_bytes += display;
-        comp3_bytes += comp3;
-        total_bytes += display + comp3;
-    }
-
-    if total_bytes == 0 {
-        total_bytes = schema.lrecl_fixed.unwrap_or(128) as u64;
-        display_bytes = (total_bytes * 80) / 100;
-        comp3_bytes = total_bytes - display_bytes;
-    }
-
-    (display_bytes, comp3_bytes, total_bytes)
-}
-
-fn parse_access_result(raw: Option<&str>) -> AccessResult {
-    match raw.unwrap_or_default().to_ascii_lowercase().as_str() {
-        "deny" | "denied" | "forbidden" | "failure" | "failed" => AccessResult::Denied,
-        "error" | "invalid" | "blocked" => AccessResult::Failed,
-        _ => AccessResult::Success,
-    }
-}
-
-fn parse_access_events(path: &Path) -> AuditResult<(Vec<AccessEvent>, Vec<String>)> {
-    let file = fs::File::open(path)?;
-    let reader = BufReader::new(file);
-    let mut events = Vec::new();
-    let mut parse_issues = Vec::new();
-
-    for (line_number, line_result) in reader.lines().enumerate() {
-        let line = match line_result {
-            Ok(value) => value,
-            Err(err) => {
-                parse_issues.push(format!(
-                    "line {}: failed to read access event line: {err}",
-                    line_number + 1
-                ));
-                continue;
-            }
-        };
-        if line.trim().is_empty() {
-            continue;
-        }
-
-        match serde_json::from_str::<RawAccessEvent>(&line) {
-            Ok(raw) => {
-                let user_id = raw
-                    .user
-                    .or(raw.user_id)
-                    .unwrap_or_else(|| "unknown-user".to_string());
-                let access_type = raw
-                    .action
-                    .or(raw.access_type)
-                    .unwrap_or_else(|| "read".to_string());
-                let resource_type = raw.resource_type.unwrap_or_else(|| "resource".to_string());
-                let resource_id = raw
-                    .resource
-                    .or(raw.resource_id)
-                    .unwrap_or_else(|| "generic".to_string());
-                let result = parse_access_result(raw.result.as_deref().or(raw.status.as_deref()));
-
-                events.push(AccessEvent {
-                    user_id,
-                    resource_type,
-                    resource_id,
-                    access_type,
-                    source_ip: raw.source_ip.or(raw.source_ip_address),
-                    user_agent: raw.user_agent,
-                    result,
-                    timestamp: None,
-                });
-            }
-            Err(err) => {
-                parse_issues.push(format!(
-                    "line {}: invalid access event JSON: {err}",
-                    line_number + 1
-                ));
-            }
-        }
-    }
-
-    Ok((events, parse_issues))
-}
-
-fn transformation_label(transformation: &TransformationType) -> &'static str {
-    match transformation {
-        TransformationType::DirectMapping => "direct_mapping",
-        TransformationType::TypeConversion => "type_conversion",
-        TransformationType::Aggregation => "aggregation",
-        TransformationType::Calculation => "calculation",
-        TransformationType::Lookup => "lookup",
-        TransformationType::Filter => "filter",
-    }
-}
-
-fn risk_level_rank(level: &RiskLevel) -> u8 {
-    match level {
-        RiskLevel::Low => 1,
-        RiskLevel::Medium => 2,
-        RiskLevel::High => 3,
-        RiskLevel::Critical => 4,
-    }
-}
-
-fn security_classification(
-    classification: Option<DataClassification>,
-) -> copybook_core::audit::context::SecurityClassification {
-    match classification {
-        Some(DataClassification::Public) => {
-            copybook_core::audit::context::SecurityClassification::Public
-        }
-        Some(DataClassification::Internal) => {
-            copybook_core::audit::context::SecurityClassification::Internal
-        }
-        Some(DataClassification::Confidential) => {
-            copybook_core::audit::context::SecurityClassification::Confidential
-        }
-        Some(DataClassification::MaterialTransaction) => {
-            copybook_core::audit::context::SecurityClassification::MaterialTransaction
-        }
-        Some(DataClassification::PHI) => copybook_core::audit::context::SecurityClassification::PHI,
-        None => copybook_core::audit::context::SecurityClassification::Internal,
-    }
-}
-
-fn is_sensitive_field_name(name: &str) -> bool {
-    let candidate = name.to_ascii_uppercase();
-    candidate.contains("SSN")
-        || candidate.contains("SOCIAL")
-        || candidate.contains("PASSWORD")
-        || candidate.contains("CREDIT")
-        || candidate.contains("CARD")
-        || candidate.contains("ACCOUNT")
-        || candidate.contains("DOB")
-        || candidate.contains("BIRTH")
-        || candidate.contains("SALARY")
-        || candidate.contains("BALANCE")
-        || candidate.contains("ROUTING")
-        || candidate.contains("TAX-ID")
-        || candidate.contains("MEDICAL")
-        || candidate.contains("PATIENT")
-        || candidate.contains("DIAGNOSIS")
-        || candidate.contains("PIN")
-        || candidate.contains("ENCRYPT")
-        || candidate.contains("SECRET")
-}
-
-fn collect_sensitive_fields(schema: &Schema) -> Vec<String> {
-    let mut sensitive = Vec::new();
-    let mut leaf_fields = Vec::new();
-    collect_leaf_fields(&schema.fields, &mut leaf_fields);
-
-    for field in leaf_fields {
-        if is_sensitive_field_name(&field.name) {
-            sensitive.push(field.name.clone());
-        }
-    }
-
-    sensitive
-}
-
-fn generate_synthetic_benchmark_data(schema: &Schema, records: usize) -> Vec<u8> {
-    let (_, _, record_size) = estimate_schema_bytes(schema);
-    let record_size = record_size.max(128);
-    let total_records = records.max(1);
-    vec![0xF0; record_size as usize * total_records]
-}
-
-fn clamp_u64(value: f64) -> u64 {
-    if !value.is_finite() || value <= 0.0 {
-        0
-    } else if value >= u64::MAX as f64 {
-        u64::MAX
-    } else {
-        value.round() as u64
-    }
-}
-
-fn report_section_from_file(path: &Path) -> AuditResult<ReportSectionFile> {
-    let content = fs::read_to_string(path)?;
-    serde_json::from_str::<ReportSectionFile>(&content).map_err(|err| err.into())
-}
+use access_log::parse_access_events;
+use health::{parse_audit_events_for_health, parse_health_events};
+use schema_analysis::{
+    collect_leaf_fields, collect_sensitive_fields, estimate_schema_bytes,
+    generate_synthetic_benchmark_data, map_field_type_name, parse_copybook_schema,
+};
+use support::{
+    bytes_to_gbps, bytes_to_mbps, clamp_u64, combine_exit_code, generate_random_suffix,
+    parse_compliance_profiles, report_section_from_file, risk_level_rank, security_classification,
+    sidecar_path, transformation_label,
+};
 
 async fn run_audit_report(
     copybook: &std::path::Path,
